+++ lib/graphql_operations/storefront/mutations/complete_checkout_token_v3.dart
@@ -17,8 +17,14 @@ mutation checkoutCompleteWithTokenizedPaymentV3($checkoutId: ID!, $payment: Toke
         amount
         currencyCode
       }
+      checkout {
+        id
+      }
       test
       id
+      nextActionUrl
+      ready
+      errorMessage
     }
   }
 }